Moved ResourceNotFound error to storage package, updated file_android.go and appropriate tests
