--- internal/driver/mobile/errors.go
@@ -1,7 +1,0 @@
-package mobile
-
-import "errors"
-
-var (
-	ErrResourceNotFound = errors.New("resource not found at URI")
-)
