+++ internal/driver/mobile/file_android.go
@@ -18,13 +18,13 @@ void closeStream(uintptr_t jni_env, uintptr_t ctx, void* stream);
 */
 import "C"
 import (
-	"errors"
 	"io"
 	"os"
 	"unsafe"
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/internal/driver/mobile/app"
+	"fyne.io/fyne/v2/storage"
 	"fyne.io/fyne/v2/storage/repository"
 )
 
@@ -91,7 +91,7 @@ func nativeFileOpen(f *fileOpen) (io.ReadCloser, error) {
 
 	ret := openStream(f.uri.String())
 	if ret == nil {
-		return nil, ErrResourceNotFound
+		return nil, storage.ErrResourceNotFound
 	}
 
 	stream := &javaStream{}
@@ -124,7 +124,7 @@ func nativeFileSave(f *fileSave) (io.WriteCloser, error) {
 
 	ret := saveStream(f.uri.String())
 	if ret == nil {
-		return nil, ErrResourceNotFound
+		return nil, storage.ErrResourceNotFound
 	}
 
 	stream := &javaStream{}
